Refactored server.js and renamed/moved client files
